Refactor keyboard monitor and related classes
